Added the Main Class
This class prompts the user for a deck file name, reads the file, and processes card information.

A unique deck ID is generated, and based on certain conditions (e.g., deck size or number of invalid cards), either a void or detailed report is produced. The non-void report includes the deck ID, total energy cost, a histogram of the energy cost distribution, and any invalid cards. The void report contains the message "VOID" without any of the information on the regular report.

This class overrode the Main class that was already in the repository, sorry Dr. Rodriguez..