Added the report class that generates the pdf document
This class generates a report for both valid decks and invalid decks.

For valid decks, a pdf document is created that contains the deck ID number, total energy cost of the deck, a histogram displaying the energy cost distribution, and a section for the invalid cards if there are any.

For invalid decks, a pdf is created that contains the "VOID" message, indicating that the deck has too many invalid cards, or the deck contains over 1000 total cards.